+++ src/Tag/Input/Range.php
@@ -97,7 +97,16 @@ final class Range extends InputTag
         return $new;
     }
 
+    /**
+     * @deprecated Use {@see addOutputAttributes()} or {@see replaceOutputAttributes()} instead. In the next major
+     * version `replaceOutputAttributes()` method will be renamed to `outputAttributes()`.
+     */
     public function outputAttributes(array $attributes): self
+    {
+        return $this->addOutputAttributes($attributes);
+    }
+
+    public function addOutputAttributes(array $attributes): self
     {
         $new = clone $this;
         $new->outputAttributes = array_merge($this->outputAttributes, $attributes);
@@ -128,7 +137,7 @@ final class Range extends InputTag
         }
 
         return "\n" . CustomTag::name($this->outputTag)
-                ->attributes($this->outputAttributes)
+                ->replaceAttributes($this->outputAttributes)
                 ->content((string) ($this->attributes['value'] ?? '-'))
                 ->id($this->outputId)
                 ->render();